% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_sample.R
\name{density_sample}
\alias{density_sample}
\title{Create a spatially thickened background sample}
\usage{
density_sample(
  x,
  dens_rast,
  samp_df,
  boundary,
  mult = 2,
  max_sample = 200,
  verbose = FALSE
)
}
\arguments{
\item{x}{spatRaster. Sample will be cell XY values from this raster.}

\item{dens_rast}{spatRaster. Density raster. This is equivalent to the
\code{weights} argument of \code{terra::spatSample()} but does not have to be at the
same resolution as \code{x}. Usually made within \code{prep_sdm()}}

\item{samp_df}{Dataframe. Usually built by \code{prep_sdm()}}

\item{boundary}{sf. Boundary within which to sample cells from \code{x}}

\item{mult}{Numeric. In each iteration try sampling \code{mult} *
\code{sum(samp_df$target)} points}

\item{max_sample}{Numeric. How many iterations to try to meet the density
targets in \code{samp_df}.}

\item{verbose}{Logical. Prints how many iterations have been attempted.}
}
\value{
Dataframe
}
\description{
Same idea as \code{terra::spatSample(weights = weight_raster)}, but can use a
density raster with a different spatial resolution. Can be much quicker than
\code{terra::spatSample()} for large areas with small resolutions.
}
\keyword{internal}
