% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_blocks.R
\name{fix_blocks}
\alias{fix_blocks}
\title{Make sure all blocks achieve min_fold_n presences}
\usage{
fix_blocks(blocks, pres, min_fold_n = 8, pres_val = 1)
}
\arguments{
\item{blocks}{Vector of block ids}

\item{pres}{Vector of presence/absence data}

\item{min_fold_n}{Minimum number of presences in a block}

\item{pres_val}{Value in \code{pres} that represent presence}
}
\value{
Vector of adjusted block ids ensuring each block achieves min_fold_n
presences
}
\description{
Make sure all blocks achieve min_fold_n presences
}
\keyword{internal}
