% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_folds.R
\name{fix_folds}
\alias{fix_folds}
\title{Make sure all folds achieve min_fold_n presences}
\usage{
fix_folds(folds, pres, min_fold_n = 8, pres_val = 1)
}
\arguments{
\item{folds}{Vector of fold ids}

\item{pres}{Vector of presence/absence data}

\item{min_fold_n}{Minimum number of presences in a fold}

\item{pres_val}{Value in \code{pres} that represent presence}
}
\value{
Vector of adjusted fold ids ensuring each fold achieves min_fold_n
presences
}
\description{
Make sure all folds achieve min_fold_n presences
}
\keyword{internal}
