% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_spatial_folds.R
\name{non_spatial_folds}
\alias{non_spatial_folds}
\title{Generate non spatial folds}
\usage{
non_spatial_folds(use_folds, data, pa_col = "pa", pres_val = 1)
}
\arguments{
\item{use_folds}{Numeric. Total number of folds}

\item{data}{Dataframe with \code{pa_col}}

\item{pa_col}{Name of column in data containing p/a values}

\item{pres_val}{Value in \code{pa_col} identifying presences}
}
\description{
Generate non spatial folds
}
\keyword{internal}
