% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/png_from_preds.R
\name{png_from_preds}
\alias{png_from_preds}
\title{Generate .png (map) files for each prediction}
\usage{
png_from_preds(
  dir,
  trim = TRUE,
  force_new = FALSE,
  include_blocks = TRUE,
  do_gc = TRUE,
  ...
)
}
\arguments{
\item{dir}{Character. Name of directory containing predicted .tif(s) to
save as .png}

\item{trim}{Logical. Trim NA values from outside (using `terra::trim()`)}

\item{force_new}{Logical. If .png file already exists, recreate it?}

\item{include_blocks}{Logical. Make a .png (map) of the blocks?}

\item{do_gc}{Logical. Run `base::rm(list = ls)` and `base::gc()` at end of
function? Useful when running SDMs for many, many taxa, especially if done in
parallel.}

\item{...}{Passed to `fs::dir_ls()`}
}
\value{
`invisible(NULL)`. Writes .png files with the same file name as any
.tif files
}
\description{
Finds any .tif files in `dir` and writes them to .png files. Includes the
retrieval and addition to the map of: various SDM metrics; and the original
presence points.
}
