% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_full_sdm.R
\name{run_full_sdm}
\alias{run_full_sdm}
\title{Run an SDM using no cross validation and previously established tune arguments}
\usage{
run_full_sdm(
  prep,
  tune,
  out_dir,
  return_val = "path",
  use_metric = "combo",
  force_new = FALSE,
  do_gc = FALSE,
  ...
)
}
\arguments{
\item{prep}{Character or named list. If character, the path to an existing
\code{prep.rds}. Otherwise, the result of a call to prep_sdm with return_val =
"object"}

\item{tune}{Character or named list. If character, the path to an existing
\code{tune.rds}. Otherwise, the result of a call to tune_sdm with return_val =
"object"}

\item{out_dir}{FALSE or character. If FALSE the result of \code{run_full_sdm()}
will be saved to a temporary folder. If character, a file 'tune.rds' will be
created at the path defined by \code{out_dir}.}

\item{return_val}{Character: "object" or "path". Both return a named list. In
the case of "path" the named list is simply list(full_run = out_dir). Will
be set to "object" if \code{out_dir} is FALSE.}

\item{use_metric}{Character. Which metric to use to find the 'best' tune
arguments from previous tuning results? Default is \code{combo}, the product of
\code{auc_po}, \code{CBI_rescale} and \code{IMAE}. \code{use_metric} must be \code{combo} or have been
used in the use_metrics argument to \code{tune_sdm()}.}

\item{force_new}{Logical. If outputs already exist, should they be remade?}

\item{do_gc}{Logical. Run \code{base::rm(list = ls)} and \code{base::gc()} at end of
function? Useful when running SDMs for many, many taxa, especially if done in
parallel.}

\item{...}{Passed to \code{tune_sdm()}}
}
\value{
If \code{return_val} is "object" a named list. If \code{return_val} is "path"
a named list \code{list(prep = out_dir)}. If \code{out_dir} is a valid path, the 'full
result' (irrespective of \code{return_val}) is also saved to
\code{fs::path(out_dir, "prep.rds")}. The 'full result' is a named list with
elements:
}
\description{
Run an SDM using no cross validation and previously established tune arguments
}
\examples{
inst/examples/predict_sdm_ex.R
}
