% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sdm_metrics}
\alias{sdm_metrics}
\title{Lookup for metrics relevant to evaluating SDMs}
\format{
A data frame with 14 rows and
6 variables:
\describe{
\item{metric}{Character. Short name of metric used in outputs}
\item{high_good}{Logical. Do high values of metric correlate with a good model?}
\item{is_thresh}{Logical. Is this metric dependent on setting a threshold?}
\item{summary_mets}{Logical. Which metrics to use in summarising?}
\item{within_mets}{Logical. Included here for compatibility with
\code{envFunc::make_metric_df()}}
\item{across_mets}{Logical. Included here for compatibility with
\code{envFunc::make_metric_df()}}
...
}
}
\usage{
sdm_metrics
}
\description{
Lookup for metrics relevant to evaluating SDMs
}
\keyword{datasets}
