% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresh_sdm.R
\name{thresh_sdm}
\alias{thresh_sdm}
\title{Threshold a previously predicted SDM}
\usage{
thresh_sdm(
  pred_file,
  this_taxa = NULL,
  threshold,
  thresh_file = NULL,
  terra_options = NULL,
  force_new = FALSE,
  do_gc = FALSE,
  check_tifs = TRUE
)
}
\arguments{
\item{pred_file}{Character. File path of predicted sdm to threshold.}

\item{this_taxa}{Character. If left as default \code{NULL} an attempt will be made
to extract a taxa name from \code{pred_file}}

\item{threshold}{Numeric. > 0 and < 1. Threshold to apply to the raster
stored in the file at \code{pred_file}. Often this value will be available within
the result of a call to \code{tune_sdm()}. e.g. \code{mod <- rio::import("tune.rds")}
and then \code{mod$e[[1]]@thresholds$max_spec_sens}}

\item{thresh_file}{Character. Name to give the output threshold. If left as
default \code{NULL}, \code{thresh_file} will be set to
\code{gsub("pred", "thresh", pred_file)}}

\item{terra_options}{Passed to \code{terra::terraOptions()}. e.g. list(memfrac =
0.6)}

\item{force_new}{Logical. If output files already exist, should they be
remade?}

\item{do_gc}{Logical. Run \code{base::rm(list = ls)} and \code{base::gc()} at end of
function? Useful to keep RAM use down when running SDMs for many, many taxa,
especially if done in parallel.}

\item{check_tifs}{Logical. Check if any output \code{.tif} files error on
\code{terra::rast()} and delete them if they do. Useful after a crash during
pred_file.}
}
\value{
Character path to threshold file, usually 'thresh.tif'. Output .tif
and .log, written to \code{out_dir}.
}
\description{
Threshold a previously predicted SDM
}
\examples{

  # setup -------
  source(fs::path("inst", "examples", "prep_sdm_ex.R")) # need 'data' object

  # data ------
  extract_thresh <- function(tune, metric = "combo", thresh_type = "max_spec_sens") {

    tune |>
      dplyr::filter(!!rlang::ensym(metric) == max(!!rlang::ensym(metric))) |>
      dplyr::pull(!!rlang::ensym(thresh_type))

  }

  data <- data |>
    dplyr::mutate(tune_mean = purrr::map(full_run, \(x) rio::import(x, trust = TRUE)$tune_mean |> dplyr::select(algo, combo, tune_args, auc_po, max_spec_sens))
                  , threshold = purrr::map_dbl(tune_mean
                                               , extract_thresh
                                               )
                  , info = basename(dirname(out_dir))
                  )


  ## thresh -------
  purrr::pwalk(list(data$pred
                    , data$threshold
                    , data$taxa
                    )
               , \(a, b, c) thresh_sdm(pred_file = a
                                       , threshold = b
                                       , this_taxa = c
                                       , thresh_file = "thresh.tif"
                                       #, force_new = TRUE
                                       )
               )

  ## visualise-------

  tifs <- data$thresh

  names <- paste0("hold_prop "
                  , data$hold_prop
                  , "; stretch "
                  , data$stretch
                  , "; new_bg "
                  , data$new_bg_test
                  )

  r <- terra::rast(tifs)
  names(r) <- names
  terra::plot(r, cex.main = 0.7, nc = 2)
}
