% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_sdm.R
\name{evaluate_sdm}
\alias{evaluate_sdm}
\title{Evaluate an SDM}
\usage{
evaluate_sdm(m, p_test, b_test, do_gc = FALSE, ...)
}
\arguments{
\item{m}{SDM result within \code{tune_sdm()}}

\item{p_test}{Presence test data generated within \code{tune_sdm()}}

\item{b_test}{Background test data generated within \code{tune_sdm()}}

\item{do_gc}{Logical. Run \code{base::rm(list = ls)} and \code{base::gc()} at end of
function? Useful when running SDMs for many, many taxa, especially if done in
parallel. Note, actually usees \code{rm(list = ls(pattern = "^[^e$]"))}.}

\item{...}{Passed to both \code{terra::predict()} and \code{predicts::pa_evaluate()}}
}
\value{
paModelEvaluation (see \code{predicts::pa_evaluate()}) with extra metrics
from \code{flexsdm::sdm_eval()}: AUC as auc_po_flexsdm; BOYCE as CBI; CBI_rescale
(CBI is -1 to 1, CBI_rescale is 0 to 1); and IMAE.
}
\description{
Returns various evaluation metrics from \code{predicts::pa_evaluate()} and
\code{flexsdm::sdm_eval()}.
}
\examples{

  out_dir <- file.path(system.file(package = "envSDM"), "examples")

  source(fs::path(out_dir, "tune_sdm_ex.R")) # make sure following prep file exists

  prep <- rio::import(fs::path(data$out_dir[[1]], "prep.rds"), trust = TRUE)

  model <- tune_sdm(prep = prep
                    , out_dir = FALSE
                    , return_val = "object"
                    , algo = "rf"
                    , trees = 500
                    , mtry = 2
                    , nodesize = 1
                    , keep_model = TRUE
                    )

  presences <- prep$testing[prep$testing$pa == 1, ]
  background <- prep$testing[prep$testing$pa == 0, ]

  evaluate_sdm(model$tune_rf$m[[1]]
               , p_test = presences
               , b_test = background
               )
}
\keyword{internal}
