% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_sdm.R
\name{evaluate_sdm}
\alias{evaluate_sdm}
\title{Evaluate an SDM}
\usage{
evaluate_sdm(m, p_test, b_test, do_gc = FALSE, ...)
}
\arguments{
\item{m}{SDM result within `tune_sdm()`}

\item{p_test}{Presence test data generated within `tune_sdm()`}

\item{b_test}{Background test data generated within `tune_sdm()`}

\item{do_gc}{Logical. Run `base::rm(list = ls)` and `base::gc()` at end of
function? Useful when running SDMs for many, many taxa, especially if done in
parallel. Note, actually usees `rm(list = ls(pattern = "^[^e$]"))`.}

\item{...}{Passed to both `terra::predict()` and `predicts::pa_evaluate()`}
}
\description{
Returns various evaluation metrics from `predicts::pa_evaluate()` and
`flexsdm::sdm_eval()`.
}
\keyword{internal}
