% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_predict_boundary.R
\name{make_predict_boundary}
\alias{make_predict_boundary}
\title{Merge polygons (or polygon files) to form a single minimum convex polygon
(mcp)}
\usage{
make_predict_boundary(
  poly_list,
  buffer_list = NULL,
  out_file = tempfile(),
  col_name = "taxa",
  col_name_val = "boundary",
  clip = NULL,
  out_crs,
  return_poly = FALSE,
  force_new = FALSE
)
}
\arguments{
\item{poly_list}{List of paths or list of sf.}

\item{buffer_list}{List, of length 1 or the same length as \code{poly_list},
specifying the distance to buffer each incoming element of \code{poly_list}.
Buffering occurs after transformation to \code{out_crs}.}

\item{out_file}{Character name of file to save}

\item{col_name}{Name of column to create in the resulting mcp}

\item{col_name_val}{Value to provide in the column in the resulting mcp}

\item{clip}{sf. Clip the resulting mcp back to this.}

\item{out_crs}{Numeric. \href{https://epsg.io/}{epsg} code}

\item{return_poly}{Logical. Return the mcp, or alternatively \code{out_file}}

\item{force_new}{Logical. If \code{out_file} exists, recreate it?}
}
\value{
If \code{return_poly}, sf, else \code{out_file}. .parquet mcp written to
\code{out_file}
}
\description{
Primary use is to create a predict boundary for \code{prep_sdm()}, merging (an
existing) minimum convex polygon with other sources of taxa distribution and
clipping to a (usually coastal) boundary. The predict boundary is then used
for generation of background points and for masking the 'full' predict (to
the extent of the environmental variables.
}
