% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/png_from_tif.R
\name{png_from_tif}
\alias{png_from_tif}
\title{Create a .png from a .tif}
\usage{
png_from_tif(
  x,
  title = NULL,
  dots = NULL,
  trim = TRUE,
  out_png = NULL,
  do_gc = FALSE
)
}
\arguments{
\item{x}{spatRaster or path to .tif}

\item{title}{Character. Title to add to the .png}

\item{dots}{sf. Usually presences. Added as points.}

\item{trim}{Logical. Run `terra::trim()` before writing to .png?}

\item{out_png}{Character. Name of .png file to save. If `NULL` will be the
same file name as `terra::sources(x)` with the file type as .png}

\item{do_gc}{Logical. Run `base::rm(list = ls)` and `base::gc()` at end of
function? Useful when running SDMs for many, many taxa, especially if done in
parallel.}
}
\description{
Create a .png from a .tif
}
