% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_sdm.R
\name{predict_sdm}
\alias{predict_sdm}
\title{Predict an SDM}
\usage{
predict_sdm(
  this_taxa = NULL,
  out_dir,
  predictors = NULL,
  is_env_pred = TRUE,
  terra_options = NULL,
  doClamp = TRUE,
  limit_to_mcp = TRUE,
  apply_thresh = TRUE,
  force_new = FALSE,
  do_gc = FALSE,
  check_tifs = TRUE,
  ...
)
}
\arguments{
\item{this_taxa}{Character. Name of taxa. Used to name outputs. If `NULL`,
this will be `basename(dirname(out_dir))`.}

\item{out_dir}{Character. Name of directory containing model to predict from
and into which results will be saved.}

\item{predictors}{Character. Vector of paths to predictor `.tif` files.}

\item{is_env_pred}{Logical. Does the naming of the directory and files in
`predictors` follow the pattern required by `envRaster::parse_env_tif()`?}

\item{terra_options}{Passed to `terra::terraOptions()`. e.g. list(memfrac = 0.6)}

\item{doClamp}{Passed to `terra::predict()` (which then passes as `...` to
`fun`). Possibly orphaned from older envSDM?}

\item{limit_to_mcp}{Logical. If `predict_boundary` exists within `prep` and
`limit_to_mcp == TRUE`, the output raster will be limted to within
`predict_boundary`.}

\item{apply_thresh}{Logical. If `TRUE`, an output raster `thresh.tif` will be
created using the threshold `mod$e[[1]]@thresholds$max_spec_sens`}

\item{force_new}{Logical. If outputs already exist, should they be remade?}

\item{do_gc}{Logical. Run `base::rm(list = ls)` and `base::gc()` at end of
function? Useful when running SDMs for many, many taxa, especially if done in
parallel.}

\item{check_tifs}{Logical. Check that any output `.tif` files error on
`terra::rast()` and delete them if they do. Useful after crash during predict.}

\item{...}{Not used.}
}
\value{
`invisible(NULL)`. Output `.tif`, log, and optional .png, written to `out_dir`
}
\description{
Predict an SDM
}
\examples{

  library("envSDM")

  out_dir <- file.path(system.file(package = "envSDM"), "examples")

  data <- file.path(system.file(package = "predicts"), "ex") |>
    fs::dir_ls(regexp = "\\\\.csv$") |>
    tibble::enframe(name = NULL, value = "path") |>
    dplyr::mutate(taxa = gsub("\\\\.csv", "", basename(path))
                  , presence = purrr::map(path, rio::import, setclass = "tibble")
                  , presence = purrr::map(presence
                                          , \(x) x |>
                                            dplyr::filter(!is.na(lat)
                                                          , !is.na(lon)
                                                          )
                                          )
                  , out_dir = fs::path(out_dir, taxa)
                  )

  # Run a single model using all the data from the 'best' tune
  purrr::pwalk(list(data$out_dir)
                 , \(a) run_full_sdm(out_dir = a
                                     , metrics_df = envSDM::sdm_metrics
                                     )
                 )

  env_dat <- system.file("ex/bio.tif", package = "predicts")

  # Predict from the best model
  purrr::pwalk(list(data$out_dir)
               , \(a) predict_sdm(out_dir = fs::path(a, "best")
                                  , predictors = env_dat
                                  , is_env_pred = FALSE
                                  , limit_to_mcp = TRUE
                                  , check_tifs = TRUE
                                  )
               )

  # make .png for each predict
  purrr::walk(data$out_dir
              , \(x) png_from_preds(x
                                    , recurse = 1
                                    , include_blocks = FALSE
                                    )
              )

  # example plots

  # masked to mcp
  purrr::walk(data$out_dir
              , \(x) fs::path(x, "best", "mask.tif") \%>\%
                terra::rast() \%>\%
                terra::trim() \%>\%
                terra::plot()
              )

  # threshold
  purrr::walk(data$out_dir
              , \(x) fs::path(x, "best", "thresh.tif") \%>\%
                terra::rast() \%>\%
                terra::trim() \%>\%
                terra::plot()
              )

}
