% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_full_sdm.R
\name{run_full_sdm}
\alias{run_full_sdm}
\title{Run an SDM using no cross validation and previously established tune arguments}
\usage{
run_full_sdm(
  this_taxa = NULL,
  out_dir,
  tune_args = NULL,
  force_new = FALSE,
  save_to = fs::path(out_dir, "best"),
  do_gc = FALSE,
  ...
)
}
\arguments{
\item{this_taxa}{Character. Name of taxa. Used to name outputs. If `NULL`,
this will be `basename(dirname(out_dir))`.}

\item{out_dir}{Character. Name of directory containing previous tunes and
into which results will be saved.}

\item{tune_args}{List with element `algo` and elements: `fc`, `rm`, `trees`,
`mtry` and `nodesize` as appropriate to the `algo.` If supplied as `NULL` an
attempt will be made to load 'best' tune from `evaluation.csv` in `out_dir`}

\item{force_new}{Logical. If outputs already exist, should they be remade?}

\item{save_to}{Character. Name of path to save results. Defaults to
`fs::path(out_dir, "best")`}

\item{do_gc}{Logical. Run `base::rm(list = ls)` and `base::gc()` at end of
function? Useful when running SDMs for many, many taxa, especially if done in
parallel.}

\item{...}{Passed to `tune_sdm()`}
}
\value{
`invisible(NULL)`. `tune.rds` saved into `save_to` directory. log
written.
}
\description{
Run an SDM using no cross validation and previously established tune arguments
}
